Defunct functions in package hereR.

The functions listed below are defunct. When possible, alternative functions with similar functionality are also mentioned. Help pages for defunct functions are available at help("-defunct").

traffic(
  aoi,
  product = "flow",
  from = NULL,
  to = NULL,
  min_jam_factor = 0,
  url_only = FALSE
)

set_auth(app_id, app_code)

unset_auth()

autocomplete(addresses, results = 5, url_only = FALSE)

set_proxy(proxy, proxyuserpwd)

traffic

For traffic(..., product = "flow"), use flow and for traffic(..., product = "incidents"), use incident.

set_auth

For set_auth, use set_key.

unset_auth

For unset_auth, use unset_key.

autocomplete

For autocomplete, use autosuggest.

set_proxy

For set_proxy, configure a global proxy for R in '~/.Renviron' instead.