Traffic Flow and Incidents in AOIs

Real-time traffic flow and incident information based on the HERE 'Traffic' API in areas of interest (AOIs). The traffic flow data contains speed ("SP") and congestion (jam factor: "JF") information. Traffic incidents contain information about location, time, duration, severity, description and other details.

traffic(
  aoi,
  product = "flow",
  from_dt = NULL,
  to_dt = NULL,
  min_jam_factor = 0,
  url_only = FALSE
)

Arguments

aoi
sf object, Areas of Interest (POIs) of geometry type POLYGON.

product
character, traffic product of the 'Traffic API'. Supported products: "flow" and "incidents".

from_dt
POSIXct object, datetime of the earliest traffic incidents (Note: Only takes effect if product is set to "incidents").

to_dt
POSIXct object, datetime of the latest traffic incidents (Note: Only takes effect if product is set to "incidents").

min_jam_factor
numeric, only retrieve flow information with a jam factor greater than the value provided (Note: Only takes effect if product is set to "flow", default = 0).

url_only
boolean, only return the generated URLs (default = FALSE)?

Value

An sf object containing the requested traffic information.

Note

Explanation of the returned traffic flow variables:

"PC": Point TMC location code.

"DE": Text description of the road.

"QD": Queuing direction. '+' or '-'. Note this is the opposite of the travel direction in the fully qualified ID, For example for location 107+03021 the QD would be ‚-‚.

"LE": Length of the stretch of road.

"TY": Type information for the given Location Referencing container. This may be a freely defined string.

"SP": Speed (based on UNITS) capped by speed limit.

"FF": The free flow speed on this stretch of the road.

"JF": The number between 0.0 and 10.0 indicating the expected quality of travel. When there is a road closure, the Jam Factor will be 10. As the number approaches 10.0 the quality of travel is getting worse. -1.0 indicates that a Jam Factor could not be calculated.

"CN": Confidence, an indication of how the speed was determined. -1.0 road closed. 1.0=100%.

References

HERE Traffic API

Flow explanation, stackoverflow

Examples

# Provide an API Key for a HERE project
set_key("<YOUR API KEY>")

# Real-time traffic flow
flow <- traffic(
  aoi = aoi[aoi$code == "LI", ],
  product = "flow",
  url_only = TRUE
)

# All traffic incidents from 2018 till end of 2019
incidents <- traffic(
  aoi = aoi[aoi$code == "LI", ],
  product = "incidents",
  from_dt = as.POSIXct("2018-01-01 00:00:00"),
  to_dt = as.POSIXct("2019-12-31 23:59:59"),
  url_only = TRUE
)