'sf'-Based Interface to the 'HERE' REST APIs

The hereR package provides an interface to the 'HERE' REST APIs:

Geocode and autocomplete addresses or reverse geocode POIs using the 'Geocoder' API;

Route directions, travel distance or time matrices and isolines using the 'Routing' API;

Request real-time traffic flow and incident information from the 'Traffic' API;

Find request public transport connections and nearby stations from the 'Public Transit' API;

Get weather forecasts, reports on current weather conditions, astronomical information and alerts at a specific location from the 'Destination Weather' API.

Locations, routes and isolines are returned as sf objects.

Application credentials

This package requires an API key for a HERE project. The key is set for the current R session and is used to authenticate in the requests to the APIs. A free login and project can be created on https://developer.here.com/. In order to obtain the API key navigate to a project of your choice in the developer portal, select 'REST: Generate APP' and then 'Create API Key'. To set the API key, please use: set_key("<YOUR API KEY>")

Functions to access the APIs

autocomplete - Get suggestions for address strings.

geocode - Get coordinates from addresses.

reverse_geocode - Get addresses or landmarks from locations.

route - Find the fastest routes between places.

route_matrix - Request a matrix of route summaries between places.

isoline - Create isochrone, isodistance or isoconsumption lines around places.

traffic - Get information about traffic jam and incidents in areas.

connection - Request public transport connections between places.

station - Find stations nearby places.

weather - Get weather observations, forecasts and alerts at places.

See also